+++ cogwheels/helpers/utils.py
@@ -31,10 +31,9 @@ class AttrReferToMethodHelper:
         self.getter_method_name = getter_method_name
 
     def __getattr__(self, name):
-        if self.settings_helper.in_defaults(name):
-            return self.get_value_via_helper_method(name)
-        raise AttributeError("{} object has no attribute '{}'".format(
-            self.settings_helper.__class__.__name__, name))
+        if not self.settings_helper.in_defaults(name):
+            self.settings_helper._raise_invalid_setting_name_error(name, error_class=AttributeError)
+        return self.get_value_via_helper_method(name)
 
     def get_value_via_helper_method(self, setting_name):
         method = getattr(self.settings_helper, self.getter_method_name)